requirements.txt
streamlit>=1.28
torch
torchvision
rfdetr>=1.4.0
supervision>=0.20.0
opencv-python
pillow
numpy
pandas
requests
